Move modal rendering and logic to EditUserLocModal component
